requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.133.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jinja2==3.1.6
kombu==5.6.2
mako==1.3.10
markupsafe==3.0.3
openrouteservice==2.3.3
prompt-toolkit==3.0.52
pydantic==2.12.5
pydantic-core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pyyaml==6.0.3
redis==7.2.1
requests==2.32.5
six==1.17.0
sqlalchemy==2.0.47
starlette==0.52.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
websockets==16.0
